check_xxx系を非推奨としてモジュールに隔離する
